Extract paired-devices helpers, fix the locale, and surface fallback-vault failures.
