+++ src/screens/Settings/PairedDevicesScreen/index.jsx
@@ -3,52 +3,21 @@ import { useEffect } from 'react'
 import { useLingui } from '@lingui/react/macro'
 import { useNavigation } from '@react-navigation/native'
 import { rawTokens, Text, useTheme } from '@tetherto/pearpass-lib-ui-kit'
-import {
-  Devices,
-  LaptopMac,
-  LaptopWindows,
-  PhoneIphone,
-  Tablet
-} from '@tetherto/pearpass-lib-ui-kit/icons'
 import { useVault } from '@tetherto/pearpass-lib-vault'
 import { StyleSheet, View } from 'react-native'
 
 import { Layout } from '../../../containers/Layout'
 import { BackScreenHeader } from '../../../containers/ScreenHeader/BackScreenHeader'
-import { unsupportedFeaturesEnabled } from '../../../utils/unsupportedFeatures'
-
-const formatDeviceDate = (dateInput) => {
-  const date = new Date(dateInput)
-  if (isNaN(date.getTime())) return ''
-  const day = date.getDate()
-  const month = date.toLocaleString('en-US', { month: 'short' })
-  const year = date.getFullYear()
-  return `${day} ${month}, ${year}`
-}
-
-const getDeviceIcon = (deviceName) => {
-  if (!deviceName) return Devices
-  const lowerName = deviceName.toLowerCase()
-  if (lowerName.startsWith('ios') || lowerName.includes('iphone'))
-    return PhoneIphone
-  if (lowerName.startsWith('android')) return PhoneIphone
-  if (lowerName.includes('ipad') || lowerName.includes('tablet')) return Tablet
-  if (
-    lowerName.includes('mac') ||
-    lowerName.includes('imac') ||
-    lowerName.includes('macbook')
-  )
-    return LaptopMac
-  if (lowerName.includes('windows')) return LaptopWindows
-  return Devices
-}
+import {
+  formatDeviceDate,
+  getDeviceIcon
+} from '../../../utils/devicePresentation'
 
 export const PairedDevicesScreen = () => {
   const { t } = useLingui()
   const navigation = useNavigation()
   const { theme } = useTheme()
   const { data, refetch: refetchVault } = useVault()
-  const showUnsupported = unsupportedFeaturesEnabled()
 
   useEffect(() => {
     refetchVault()
@@ -88,7 +57,7 @@ export const PairedDevicesScreen = () => {
 
             return (
               <View
-                key={device.name + index}
+                key={device.id ?? `${device.name}-${index}`}
                 style={[
                   styles.deviceItem,
                   !isLast && {
@@ -118,15 +87,6 @@ export const PairedDevicesScreen = () => {
                     {getDeviceDisplayName(device?.name)}
                   </Text>
                   <View style={styles.deviceDates}>
-                    {showUnsupported && (
-                      <Text
-                        variant="caption"
-                        color={theme.colors.colorTextSecondary}
-                        numberOfLines={1}
-                      >
-                        {t`Last used on ${formatDeviceDate(device.createdAt)}`}
-                      </Text>
-                    )}
                     {device.createdAt && (
                       <Text
                         variant="caption"
