+++ src/types/pearpass-lib-vault.d.ts
@@ -134,7 +134,7 @@ declare module '@tetherto/pearpass-lib-vault' {
         hashedPassword?: string
       }
     ) => Promise<unknown>
-    addDevice: (device: unknown) => Promise<void>
+    addDevice: () => Promise<void>
     isVaultProtected: (vaultId: string) => Promise<boolean>
     deleteVaultLocal: (vaultId: string) => Promise<unknown[]>
   }
@@ -247,7 +247,11 @@ declare module '@tetherto/pearpass-lib-vault' {
   ): Promise<Array<{ recordId: string; code: string; timeRemaining?: number }>>
   export function generateHotpNext(recordId: string): Promise<{ code: string } | null>
   export function closeAllInstances(): Promise<void>
-  export function setPearpassVaultClient(client: unknown): void
+  export function setPearpassVaultClient(
+    client: unknown,
+    options?: { currentDeviceName?: string }
+  ): void
+  export function setCurrentDeviceName(name: string | null): void
   export function setStoragePath(path: string): void
   export function authoriseCurrentProtectedVault(params: unknown): Promise<unknown>
   export function getVaultById(vaultId: string): Promise<Record<string, any>>
